Add the body size parameter to rubus messages
The body-length parameter contains the length of the body in bytes. The
body is separated from the header with an empty line ( one line feeder
is a part of the header and another one is a separator ). The receiving
side use the parameter and the empty line to determine the total size
of the application message.