Retrieve the rubus message from the socket in a proper way
The methods used to constantly read bytes while they are available.
After that the RequestHandler thread was locked because there were no
timeout set. And RubusClient.send could return a RubusResponse instance
containing more bytes than the original response and threw an
unnecessary timeout exception.

The current implementations read only the necessary amount of bytes
and exit if there is no byte available or the message is corrupted
in a certain way by throwing an exception.