Fix video and audio pieces retrieval
The bug principal is the same in both methods. The for loop
initialized a variable with the initial value of pieceIndex and was
supposed to cycle number times. But it held true only if pieceIndex
had a value of zero, otherwise it cycled less than number times or
didn't cycle at all returning a null-filled array.