Extend PlayerInterface
The isBuffering method was added so the client could determine whether
the player has actually data to play or not.

Before that the client, FetchController in particular, assumed it by
looking at getBuffer size. The downside of this approach is that
the client had to assumed it based on the implementation. getBuffer
should return the data that is in the line to play, but it actually
contained the data that was already played.