Improve the overall frame pacing
A new frame was drawn only if it was x or more nanoseconds since
the previous frame had been drawn. If the player is not updated at
consistent rate or the update rate isn't a fraction of x ( like when
player update rate is 60hz but the frame pace is 24hz ) the actual frame
pace is constantly getting larger the more frames are drawn.

To make it as close to x as possible the deviation variable is
introduced. It accumulates the difference between x and the actual
frame pace. So when the next pace is about to be drawn the actual pacing
is going to be x + deviation. The deviation variable is being constantly
updated so the frame actual pacing will be ~x consistently regardless
of the video duration.