Fix the video seeking bug
If a new video position was set within the buffered range, instead of
skipping n pieces it left n pieces in the buffer. For example if there
were 10 pieces in the buffer and the user moved the video position from
0 to 2, it removed all but the 2 last pieces from the buffer.
It also didn't check if the video is already playing in which case only
n-1 pieces need to be removed.