Fix the accidental thread deadlock
The SocketManager thread is suspended if the socket queue is empty.
Although the add method would wake up the thread if the queue
was empty before the invocation, it failed to do so under certain
conditions. So instead of checking if the queue is empty the add method
checks if the SocketManager thread is waiting.