Fix the bug that miscalculates the number of open connections
The SocketManager.size() method returned the size of the underlying
socket queue, but most of the time the sockets are pulled out of
the queue and passed to RequestHandler. This means
the SocketManager.size() returned less opened connections that there
actually were. The solution is to keep the number of opened connection
in a variable and increment the value whenever a new socket is being
added and decrement the value whenever an opened socket is being closed.