Get rid of the background fetching thread interruption
The intent behind the initial design was the simplicity it provided. If
the state of player has changed before the background fetch is finished
is usually due to changing of the video position by the user. And if
the position has changed there is a chance the background fetch is not
relevant anymore.
It, however, has tons of shortcomings:
    1. It implies that the fetching takes less than one second, which
       is not always the case.
    2. The background fetch may still be relevant if the position change
       is positive and less than the data in the video buffer.
    3. The thread interruption happened only during the setting of
       the video buffer; the network communication with the server
       continues to run ( and it's impossible to interrupt if the sever
       has already sent the message and the background fetch is waiting
       to receive it ). That means there could be two or more threads,
       both waiting to receive the response. So there is a chance a new
       background fetch threads receiving the response sent to the old
       one.

The solution is to check if the data retrieved by the already running
background fetch is still relevant. And if it's not wait until it
finishes so it can rerun the process to fetch the relevant data.

The notification that the data is not relevant and it needs to rerun is
set by the client.