Replace Playback with Media when generalization is implied
Media is a more general world when referred to some content like video,
music, etc. Whereas playback usually implies something concrete. So,
when, for example, the client requests what videos or music are on
the server it refers to them as media. But when referring to the actual
data of the media playback pieces is more appropriate because they are
used to actually play the specific type of media ( music, video, etc.).