Close a socket when the background fetching gets irrelevant
In the previous versions RubusClient as well as FetchController, which
creates a RubusClient instance, lacked the control over the socket it
used. Now RubusClient can create and close sockets on demand meaning
FetchController now has indirect control over it.

FetchController uses the same algorithm to determine if the currently
fetching thread and the data it's fetching is still relevant. It
basically computes what index of the next playback piece it needs to
fetch and compare it to the fetching thread's one. If they are the same
the thread is still relevant; if they are different it is not relevant.

So if the currently running background thread is irrelevant,
FetchController now simply interrupts it and closes the socket via
RubusClient.close() then it starts a new background thread and
the RubusClient instance opens a new socket automatically. By closing
the socket it's guaranteed that the new background thread won't receive
the response intended for the old background thread by accident,
because now they would use different sockets.