Fix the last playback piece skip bug
When there is only one playback piece left in the buffer, player takes
it out, starts decoding, and set the buffering variable to true. But
that triggered the condition on the line 291, which is intended to
trigger only if there is no pieces no play.
The solution is to change the condition, so it compares the decoder of
the current second to null. After the last piece is played the decoder
is set to null, so the condition triggers, when there is actually no
more data to play.