Change the constructor arguments of RubusServer
RubusServer doesn't initialize the SocketManager and RubusServerSocket
objects and instead delegates their construction to the client. It is
done to lose the coupling between RubusServer and these 2 objects.

P.S.
RubusServer used RubusSockets to initialize the RubusServerSocket
object, and it may seem that RubusServer and RubusServerSocket were not
tightly coupled. This, however, is false because RubusServer still had
to pass arguments to the factory method which, in turn, involves some
assumptions regarding the way RubusServerSocket is constructed.