Remove redundant business logic from SocketManager
Initially, SocketManager was designed in the way that it extended
the Thread class and overrode the run method. Inside the run method it
pooled the sockets from the concurrent queue, which had initially been
added in there via add() or keepConnection(). This is absolutely
unnecessary because SocketManager unconditionally wrapped sockets
inside RequestHandlers and submitted them to the underlying executor
service. There were no decision-making nor SocketManager exposed
the sockets to other classes. So, in the chain
queue -> executor service -> queue -> executor service the queue is
just a middle man which can be removed.
In addition to that, SocketManager now reuses RequestHandlers instead
of constantly constructing them before submitting them to the executor
manager.