Remove the media progress retrieval from MainFrame
The media progress used to get retrieved twice for a media the user
choosing to play: the first time when it's displayed on the media list
with the label indicating the progress; and the second time in
MainFrame.play, where MainFrame initiates a PlayerInterface instance
and passes the progress to this instance. It's a clear duplication of
the code, so now the progress is simply passed to MainFrame.play as
a parameter by MediaListPanel.