Overhaul the rendering process in Player
Player now uses VideoDecoder instead of BMPDecoder and so its behavior
changed appropriately. Another change of rendering is that it's able
to downscale or upscale the rendered frames. That makes Player and
MainFrame independent of the size of the video, so MainFrame just
stretches or shrinks Player to the available size.