Fix the audio desynchronization bug
All audio pieces are meant to be played in exactly one second. But due
to the bug the duration was constantly exceeded by ~3 ms. Although it
may seem insignificant the surplus accumulated over the time and in just
a few minutes the audio lag was noticeable. The specification of
the javax.sound.sampled package is quite outdated and its implementation
seems to be buggy, so some methods return values that don't make sense,
which makes the workaround of this bug is rather quirky.

The original implementation was to write some audio frames to
audioOutput and suspend the current thread so a single audio clip
would be played in one second. Unfortunately, Thread.sleep is not quite
precise and, in addition to this, writing audio frames takes some time
as well, which makes me to believe the delay was caused by
the combination of these two factors.

The specification of SourceDataLine says that the writing is blocked
if the buffer is full, which is the solution to the bug. So the idea
is to make the buffer much smaller ( of the size to fit the number of
frames that need to be played per update ) so it would be always full.
The frames that needs to be written in the next update would have to
wait for the entire buffer to get empty. It makes the playback of
the audio clip ~1 second and the desynchronization bug disappears.

Unfortunately, it led to another bug that caused audio desynchronization
after the video was paused and resumed, but it was easily fixed by
lowering the update time.