Remove several Media's members
The removed method of the MediaInterface were used to access
the information that can be inferred from the metadata of the media
itself. The only members that were actually used are getVideoContainer()
and getAudioContainer() to access the extension of the file. So
the underlying media resources are now extensionless to safely remove
the two methods.